fix: use BigDecimal for product price to avoid floating-point errors
Change price fields from Double to BigDecimal in Product and ProductDTO,
and rework discount calculation in ProductService to use BigDecimal with
setScale(2, HALF_UP), fixing precision errors like 8.406585 appearing
instead of the expected rounded value

+++ product_api/src/main/java/com/saraprojects/product_api/dto/ProductDTO.java
@@ -23,9 +23,9 @@ public class ProductDTO {
 
     @NotNull
     @Positive
-    private Double price;
+    private BigDecimal price;
 
-    private Double originalPrice;
+    private BigDecimal originalPrice;
 
     private BigDecimal discountPercentage;
